Adding Preemptive authentication
HttpClient does not support preemptive authentication out of the box resulting in 401 errors being thrown at the API servers. Every request coming from the clients using the SDK is resulting in two calls. A 401 and then a subsequent 202 (when the proxy is now passing the credentials).

This code change is targeted to add the preemptive authentication to avoid the duplicate 401